Build 2.0.1378 - Fixed #15558,#15548 Applied 2 Pathes
git-svn-id: https://avalondock.svn.codeplex.com/svn@95419 4ef1bfeb-0a34-4d02-9305-240e5b07a168

+++ Version2.0/AvalonDock/AvalonDock/Win32Helper.cs
@@ -172,9 +172,9 @@ namespace AvalonDock
         internal const int HT_CAPTION = 0x2;
 
         [DllImportAttribute("user32.dll")]
-        internal static extern int SendMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
+        internal static extern int SendMessage(IntPtr hWnd, int Msg, IntPtr wParam, IntPtr lParam);
         [DllImportAttribute("user32.dll")]
-        internal static extern int PostMessage(IntPtr hWnd, int Msg, int wParam, int lParam);
+        internal static extern int PostMessage(IntPtr hWnd, int Msg, IntPtr wParam, IntPtr lParam);
 
 
         [DllImport("user32.dll")]
